+++ src/Field/Radio.php
@@ -7,6 +7,8 @@
 
 namespace Del\Form\Field;
 
+use Del\Form\Renderer\Field\RadioRender;
+
 class Radio extends FieldAbstract
 {
     /**
@@ -21,7 +23,6 @@ class Radio extends FieldAbstract
     {
         $this->setAttribute('type', 'radio');
         $this->setAttribute('class', 'form-control');
+        $this->setRenderer(new RadioRender());
     }
-
-
 }
