+++ src/Renderer/Field/TextRender.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * User: delboy1978uk
+ * Date: 04/12/2016
+ * Time: 21:08
+ */
+
+namespace Del\Form\Renderer\Field;
+
+class TextRender extends FieldRender implements FieldRendererInterface
+{
+
+}
